Flagicon fix
Although it was a cool easter egg, I noticed that Gareth Barry was often 
changed to ' flagicon', so it would be possible to figure out it was 
Barry, enter barry, and be wrong. then realise it is flagicon, type 
flagicon, and be wrong (as there is a space prior to the word).

> if statement added in SetPlayer function to see if the player randName 
state is ' flagicon' and if so, set to Gareth Barry. (Clicked through 
around 80 times, only ever seemed to be Barry affected.

> Removed a lot of console logs that filled the console in browser